Renamed helper and context modules
